// src/post/dto/search-post-dto.ts
// Post - DTO - это объект, который используется для передачи данных
// между разными частями вашего приложения. Обычно он применяется
// для передачи данных между клиентом и сервером
// или между разными слоями вашего серверного кода.
// В вашем случае, SearchPostDto - это объект,
// который определяет параметры поиска для постов. Он имеет поля,
// которые могут быть заполнены, например, клиентом при запросе на сервер
// для поиска конкретных постов по определенным критериям,
// таким как title, body, views, limit, take, tag.
// тоесть мы еще описываем параметры ,которые будем передавать на бек с фронта
// для получения нужных нам данныхи и сортировки их

export class SearchPostDto {
  id?: number;
  title?: string;
  body?: string;
  views?: 'DESC' | 'ASC';
  limit?: number;
  take?: number;
  tag?: string;
}